Made component to hold header and flow logic
