+++ client/src/App.jsx
@@ -1,14 +1,12 @@
 import React from 'react';
-import Flow from './components/Flow/Flow.jsx';
-import Header from './components/Header/Header.jsx';
+import MainContainer from './components/main-container/main-container.jsx';
 
 import './App.scss';
 
 export default function App() {
   return (
     <div className="app">
-      <Header />
-      <Flow />
+      <MainContainer />
     </div>
   );
 }
